ensure options in run-rebuild,test.sh

+++ run-rebuild.sh
@@ -5,39 +5,69 @@ rclcppRoot=${testRoot}/rclcpp_node
 rclexRoot=${testRoot}/rclex_node
 
 
-# Rebuild Rclcpp node
-cd $rclcppRoot
-if [ "$(uname)" == 'Darwin' ];
+if [ $# -eq 0 ];
 then
-    rm -rf build install log
+    build_rclcpp=1
+    build_rclex=1
 else
-    sudo rm -rf build install log
+    if [ "$1" == "rclcpp" ];
+    then
+        build_rclcpp=1
+        build_rclex=0
+    elif [ "$1" == "rclex" ];
+    then
+        build_rclcpp=0
+        build_rclex=1
+    else
+        echo "ERROR: invalid option: ${1}"
+        exit 1
+    fi
 fi
 
-colcon build
-if [ $? -ne 0 ]; then
-    echo "ERROR: \`colcon build\` for Rclcpp failed: $result"
-    exit $result
+
+# Rebuild Rclcpp node
+if [ ${build_rclcpp} -eq 1 ];
+then
+    echo "INFO: building rclcpp node in ${rclcppRoot}"
+    cd $rclcppRoot
+    if [ "$(uname)" == 'Darwin' ];
+    then
+        rm -rf build install log
+    else
+        sudo rm -rf build install log
+    fi
+
+    colcon build
+    if [ $? -ne 0 ]; then
+        echo "ERROR: \`colcon build\` for Rclcpp failed: $result"
+        exit $result
+    fi
 fi
 
 
 # Rebuild Rclex node
-cd $rclexRoot
-if [ "$(uname)" == 'Darwin' ];
+if [ ${build_rclex} -eq 1 ];
 then
-    rm -rf _build deps
-else
-    sudo rm -rf _build deps
-fi
-
-mix deps.get
-if [ $? -ne 0 ]; then
-    echo "ERROR: \`mix deps.get\` for Rclex failed: $result"
-    exit $result
-fi
-
-mix compile
-if [ $? -ne 0 ]; then
-    echo "ERROR: \`mix compile\` for Rclex failed: $result"
-    exit $result
+    echo "INFO: building rclcpp node in ${rclexRoot}"
+    cd $rclexRoot
+    if [ "$(uname)" == 'Darwin' ];
+    then
+        rm -rf _build deps
+    else
+        sudo rm -rf _build deps
+    fi
+    
+    mix deps.get
+    if [ $? -ne 0 ];
+    then
+        echo "ERROR: \`mix deps.get\` for Rclex failed: $result"
+        exit $result
+    fi
+    
+    mix compile
+    if [ $? -ne 0 ];
+    then
+        echo "ERROR: \`mix compile\` for Rclex failed: $result"
+        exit $result
+    fi
 fi
